# Conflicts:
#	portfolio/build.gradle
#	portfolio/src/test/java/nz/ac/canterbury/seng302/portfolio/controller/ViewAccountControllerTest.java

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/ProjectController.java
@@ -3,12 +3,16 @@ package nz.ac.canterbury.seng302.portfolio.controller;
 import nz.ac.canterbury.seng302.portfolio.model.contract.BaseProjectContract;
 import nz.ac.canterbury.seng302.portfolio.model.contract.ProjectContract;
 import nz.ac.canterbury.seng302.portfolio.service.ProjectService;
+import nz.ac.canterbury.seng302.portfolio.service.RolesService;
 import nz.ac.canterbury.seng302.portfolio.service.ValidationService;
+import nz.ac.canterbury.seng302.shared.identityprovider.AuthState;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import java.util.ArrayList;
+import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Objects;
 
@@ -26,18 +30,25 @@ public class ProjectController {
     @Autowired
     private ValidationService validationService;
 
+    @Autowired
+    private RolesService rolesService;
+
     /**
      * This method will be invoked when API receives a GET request, and will produce a list of all the projects.
      * @return List of projects converted into project contract (JSON) type.
      */
     @GetMapping(value = "/projects", produces = "application/json")
-    public ResponseEntity<?> getAll() {
-        try{
-            ArrayList<ProjectContract> projects = projectService.allProjects();
-            return ResponseEntity.ok(projects);
-        }
-        catch (NoSuchElementException error) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+    public ResponseEntity<?> getAll(@AuthenticationPrincipal AuthState principal) {
+        ArrayList<String> roles = rolesService.getRolesByToken(principal);
+        if (roles.contains("TEACHER") || roles.contains("COORDINATOR")) {
+            try {
+                ArrayList<ProjectContract> projects = projectService.allProjects();
+                return ResponseEntity.ok(projects);
+            } catch (NoSuchElementException error) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
     }
 
@@ -63,21 +74,25 @@ public class ProjectController {
      * @return a project contract (JSON) type of the newly created project.
      */
     @PostMapping(value = "/projects", produces = "application/json")
-    public ResponseEntity<?> addNewProject(@RequestBody BaseProjectContract newProject) {
-        try {
-            var errorMessage = validationService.checkAddProject(newProject);
+    public ResponseEntity<?> addNewProject(@AuthenticationPrincipal AuthState principal, @RequestBody BaseProjectContract newProject) {
+        ArrayList<String> roles = rolesService.getRolesByToken(principal);
+        if (roles.contains("TEACHER") || roles.contains("COORDINATOR")) {
+            try {
+                var errorMessage = validationService.checkAddProject(newProject);
 
-            if (!errorMessage.equals("Okay")) {
-                if (errorMessage.equals("Project ID does not exist")) {
-                    return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorMessage);
+                if (!errorMessage.equals("Okay")) {
+                    if (errorMessage.equals("Project ID does not exist")) {
+                        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorMessage);
+                    }
+                    return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorMessage);
                 }
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorMessage);
+                var project = projectService.create(newProject);
+                return ResponseEntity.ok(project);
+            } catch (Exception error) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
             }
-            var project = projectService.create(newProject);
-            return ResponseEntity.ok(project);
-        }
-        catch (Exception error) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
     }
 
@@ -87,16 +102,19 @@ public class ProjectController {
      * @return a project contract (JSON) type of the project.
      */
     @DeleteMapping(value = "/projects/{id}", produces = "application/json")
-    public ResponseEntity<?> removeProject(@PathVariable String id) {
-        try{
-            projectService.delete(id);
-            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-        }
-        catch(NoSuchElementException error) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
-        catch (Exception error) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+    public ResponseEntity<?> removeProject(@AuthenticationPrincipal AuthState principal, @PathVariable String id) {
+        ArrayList<String> roles = rolesService.getRolesByToken(principal);
+        if (roles.contains("TEACHER") || roles.contains("COORDINATOR")) {
+            try {
+                projectService.delete(id);
+                return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+            } catch (NoSuchElementException error) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            } catch (Exception error) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
     }
 
@@ -106,21 +124,25 @@ public class ProjectController {
      * @return a project contract (JSON) type of the project.
      */
     @PutMapping(value = "/projects/{id}", produces = "application/json")
-    public ResponseEntity<?> updateProject(@RequestBody ProjectContract updatedProject, @PathVariable String id) {
-        try {
-            var errorMessage = validationService.checkUpdateProject(id, updatedProject);
+    public ResponseEntity<?> updateProject(@AuthenticationPrincipal AuthState principal, @RequestBody ProjectContract updatedProject, @PathVariable String id) {
+        ArrayList<String> roles = rolesService.getRolesByToken(principal);
+        if (roles.contains("TEACHER") || roles.contains("COORDINATOR")) {
+            try {
+                var errorMessage = validationService.checkUpdateProject(id, updatedProject);
 
-            if (!errorMessage.equals("Okay")) {
-                if (errorMessage.equals("Project ID does not exist")) {
-                    return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorMessage);
+                if (!errorMessage.equals("Okay")) {
+                    if (errorMessage.equals("Project ID does not exist")) {
+                        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorMessage);
+                    }
+                    return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorMessage);
                 }
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorMessage);
+                projectService.update(updatedProject, id);
+                return ResponseEntity.ok("");
+            } catch (NoSuchElementException error) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
             }
-            projectService.update(updatedProject, id);
-            return ResponseEntity.ok("");
-        }
-        catch (NoSuchElementException error) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
     }
 }
